feat: added dtos

+++ backend/tienda-api/src/main/java/com/delogica/tienda_api/domain/Order.java
@@ -1,6 +1,5 @@
 package com.delogica.tienda_api.domain;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
@@ -38,7 +37,7 @@ public class Order {
     private LocalDateTime orderDate;
 
     @Column(precision = 10, scale = 2)
-    private BigDecimal total;
+    private Double total;
 
     @OneToMany(mappedBy = "order", fetch = FetchType.LAZY, cascade = CascadeType.ALL)
     private List<OrderItem> items = new ArrayList<>();
